Fix browser caching in Main Donation Form
The main donation form makes use of `input` elements with `type="checkbox"`. It
is common for browsers such as Mozilla Firefox to cache the state of a checkbox.
This would lead to an issue where the user is able to toggle the `Anyone in
need` item, refresh the page and the other elements will no longer be `disabled`
leading to incorrect output when the Stripe donation page was shown.

Additionally I have renamed the `showAll` value to `isAnyoneInNeedToggled` this
name is longer but more representative of the actual value held.