Delete .github/PULL_REQUEST_TEMPLATE directory
As per #194 there appears to be no reason to have this as it cannot be used in the same way that issue templates can be used. Please see the relevant self-learning discussions within the issue if you would like to read the specific details in what I was attempting and what has been shown through a trial and error process.